Housekeeping

+++ app/modules/dbutils/db_credentials.py
@@ -59,6 +59,9 @@ def del_credentials(credentials_id: int) -> bool:
     return:
         bool
     """
+    if not isinstance(credentials_id, int) or credentials_id is None:
+        logger.info(f"Delete credentials error, credentials_id must be an integer")
+        return False
     try:
         Credentials.query.filter_by(id=int(credentials_id)).delete()
         db.session.commit()
@@ -139,10 +142,13 @@ def get_all_credentials() -> list:
     ]
 
 
-def get_credentials(credentials_id: int) -> dict:
+def get_credentials(credentials_id: int) -> dict or None:
     """
     This function return credentials
     """
+    if not isinstance(credentials_id, int) or credentials_id is None:
+        logger.info(f"Getting credentials error, credentials_id must be an integer")
+        return None
     credentials = (
         Credentials.query.with_entities(
             Credentials.id,
@@ -163,35 +169,37 @@ def get_credentials(credentials_id: int) -> dict:
     }
 
 
-def get_allowed_credentials(user_id: int) -> list:
+def get_allowed_credentials(user_id: int) -> list or None:
     """
     This function needs to get allowed credentials for a user
     """
-    if isinstance(user_id, int) and user_id is not None:
-        try:
-            slq_request = text(
-                "select "
-                "credentials.id, "
-                "credentials.credentials_name, "
-                "credentials.credentials_username, "
-                "credentials.user_group_id "
-                "from credentials "
-                "left join group_permission on group_permission.user_group_id = credentials.user_group_id  "
-                "where group_permission.user_id = :user_id"
-            )
-            parameters = {"user_id": user_id}
-            credentials_data = db.session.execute(slq_request, parameters).fetchall()
-            return [
-                {
-                    "html_element_id": html_element_id,
-                    "credentials_id": i["id"],
-                    "credentials_name": i["credentials_name"],
-                    "credentials_username": i["credentials_username"],
-                }
-                for html_element_id, i in enumerate(credentials_data)
-            ]
+    if not isinstance(user_id, int) or user_id is None:
+        logger.info(f"Getting credentials error, user_id must be an integer")
+        return None
+    try:
+        slq_request = text(
+            "select "
+            "credentials.id, "
+            "credentials.credentials_name, "
+            "credentials.credentials_username, "
+            "credentials.user_group_id "
+            "from credentials "
+            "left join group_permission on group_permission.user_group_id = credentials.user_group_id  "
+            "where group_permission.user_id = :user_id"
+        )
+        parameters = {"user_id": user_id}
+        credentials_data = db.session.execute(slq_request, parameters).fetchall()
+        return [
+            {
+                "html_element_id": html_element_id,
+                "credentials_id": i["id"],
+                "credentials_name": i["credentials_name"],
+                "credentials_username": i["credentials_username"],
+            }
+            for html_element_id, i in enumerate(credentials_data)
+        ]
 
-        except Exception as get_sql_error:
-            # If an error occurs as a result of writing to the DB,
-            # then rollback the DB and write a message to the log
-            logger.info(f"getting allowed credentials error {get_sql_error}")
+    except Exception as get_sql_error:
+        # If an error occurs as a result of writing to the DB,
+        # then rollback the DB and write a message to the log
+        logger.info(f"getting allowed credentials error {get_sql_error}")
